+++ ndk/src/events/index.ts
@@ -10,7 +10,7 @@ import type { NDKSigner } from "../signers/index.js";
 import type { NDKFilter } from "../subscription/index.js";
 import type { NDKUser } from "../user/index.js";
 import Zap from "../zap/index.js";
-import { type ContentTag, generateContentTags } from "./content-tagger.js";
+import { type ContentTag, generateContentTags, mergeTags } from "./content-tagger.js";
 import { isEphemeral, isParamReplaceable, isReplaceable } from "./kind.js";
 import { NDKKind } from "./kinds/index.js";
 import { decrypt, encrypt } from "./nip04.js";
@@ -116,18 +116,12 @@ export class NDKEvent extends EventEmitter {
      */
     public tag(event: NDKEvent, marker?: string): void;
     public tag(userOrEvent: NDKUser | NDKEvent, marker?: string): void {
-        const tags = userOrEvent.referenceTags();
-        if (marker) tags[0].push(marker);
-        this.tags.push(...tags);
+        const skipAuthorTag = userOrEvent?.pubkey === this.pubkey;
+        const tags = userOrEvent.referenceTags(marker, skipAuthorTag);
+
+        this.tags = mergeTags(this.tags, tags);
 
         if (userOrEvent instanceof NDKEvent) {
-            const tagEventAuthor = userOrEvent.author;
-
-            // If event author is not the same as the user signing this event, tag the author
-            if (tagEventAuthor && this.pubkey !== tagEventAuthor.hexpubkey) {
-                this.tag(tagEventAuthor);
-            }
-
             // tag p-tags in the event if they are not the same as the user signing this event
             for (const pTag of userOrEvent.getMatchingTags("p")) {
                 if (pTag[1] === this.pubkey) continue;
@@ -361,9 +355,11 @@ export class NDKEvent extends EventEmitter {
         return `${this.kind}:${this.pubkey}:${dTagId}`;
     }
 
-    /** @deprecated Use referenceTags instead. */
     /**
-     * Get the tag that can be used to reference this event from another event
+     * Get the tag that can be used to reference this event from another event.
+     *
+     * Consider using referenceTags() instead (unless you have a good reason to use this)
+     *
      * @example
      *     event = new NDKEvent(ndk, { kind: 30000, pubkey: 'pubkey', tags: [ ["d", "d-code"] ] });
      *     event.tagReference(); // ["a", "30000:pubkey:d-code"]
@@ -372,13 +368,28 @@ export class NDKEvent extends EventEmitter {
      *     event.tagReference(); // ["e", "eventid"]
      * @returns {NDKTag} The NDKTag object referencing this event
      */
-    tagReference(): NDKTag {
+    tagReference(marker?: string): NDKTag {
+        let tag: NDKTag;
+
         // NIP-33
         if (this.isParamReplaceable()) {
-            return ["a", this.tagAddress()];
+            tag = ["a", this.tagAddress()];
+        } else {
+            tag = ["e", this.tagId()];
         }
 
-        return ["e", this.tagId()];
+
+        if (this.relay) {
+            tag.push(this.relay.url);
+        } else {
+            tag.push("");
+        }
+
+        if (marker) {
+            tag.push(marker);
+        }
+
+        return tag;
     }
 
     /**
@@ -392,7 +403,7 @@ export class NDKEvent extends EventEmitter {
      *     event.referenceTags(); // [["e", "parent-id"]]
      * @returns {NDKTag} The NDKTag object referencing this event
      */
-    referenceTags(marker?: string): NDKTag[] {
+    referenceTags(marker?: string, skipAuthorTag?: boolean): NDKTag[] {
         let tags: NDKTag[] = [];
 
         // NIP-33
@@ -405,11 +416,25 @@ export class NDKEvent extends EventEmitter {
             tags = [["e", this.id]];
         }
 
+        // Add the relay url to all tags
+        if (this.relay?.url) {
+            tags = tags.map((tag) => {
+                tag.push(this.relay?.url!);
+                return tag;
+            });
+        } else if (marker) {
+            tags = tags.map((tag) => {
+                tag.push("");
+                return tag;
+            });
+        }
+
         if (marker) {
             tags.forEach((tag) => tag.push(marker)); // Add the marker to both "a" and "e" tags
         }
 
-        tags.push(...this.author.referenceTags());
+        if (!skipAuthorTag)
+            tags.push(...this.author.referenceTags());
 
         return tags;
     }