+++ ndk/src/signers/nip46/backend/get-public-key.ts
@@ -3,9 +3,10 @@ import type { IEventHandlingStrategy, NDKNip46Backend } from "./index.js";
 export default class GetPublicKeyHandlingStrategy implements IEventHandlingStrategy {
     async handle(
         backend: NDKNip46Backend,
+        id: string,
         remotePubkey: string,
         params: string[]
     ): Promise<string | undefined> {
-        return backend.localUser?.hexpubkey;
+        return backend.localUser?.pubkey;
     }
 }