+++ ndk-svelte/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @nostr-dev-kit/ndk-svelte
 
+## 2.0.7
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nostr-dev-kit/ndk@2.1.0
+
+## 2.0.6
+
+### Patch Changes
+
+-   Updated dependencies
+-   Updated dependencies
+    -   @nostr-dev-kit/ndk@2.0.6
+
 ## 2.0.5
 
 ### Patch Changes