+++ ndk/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @nostr-dev-kit/ndk
 
+## 2.1.0
+
+### Minor Changes
+
+-   New NIP-46 implementation with auth_url flow
+
+## 2.0.6
+
+### Patch Changes
+
+-   Add relay hint to tagging via tagReference
+-   fix tagging and add more tests
+
 ## 2.0.5
 
 ### Patch Changes